Split async notification queue tests
