+++ .github/swift-file-length-budget.tsv
@@ -1,10 +1,10 @@
 # cmux-owned Swift file length budget.
 # Format: max_lines<TAB>relative path
 # Reduce counts as files shrink. CI fails if tracked files exceed this budget.
-18705	CLI/cmux.swift
-16364	Sources/TerminalController.swift
+18755	CLI/cmux.swift
+16411	Sources/TerminalController.swift
 15762	Sources/ContentView.swift
-13748	Sources/AppDelegate.swift
+13778	Sources/AppDelegate.swift
 13702	Sources/Workspace.swift
 13208	Sources/GhosttyTerminalView.swift
 10549	Sources/Panels/BrowserPanel.swift
@@ -37,10 +37,10 @@
 1563	cmuxTests/WindowAndDragTests.swift
 1517	cmuxUITests/BrowserPaneNavigationKeybindUITests.swift
 1502	Sources/Update/UpdateTitlebarAccessory.swift
-1423	Sources/TerminalNotificationStore.swift
+1461	Sources/TerminalNotificationStore.swift
 1365	Sources/Feed/FeedButtonStyleDebugWindowController.swift
 1292	cmuxUITests/SidebarHelpMenuUITests.swift
-1255	cmuxTests/NotificationAndMenuBarTests.swift
+1265	cmuxTests/NotificationAndMenuBarTests.swift
 1215	cmuxUITests/MenuKeyEquivalentRoutingUITests.swift
 1187	cmuxUITests/MultiWindowNotificationsUITests.swift
 1165	Sources/Auth/AuthManager.swift
@@ -58,7 +58,7 @@
 752	cmuxUITests/CloseWorkspaceCmdDUITests.swift
 719	cmuxUITests/BrowserOmnibarSuggestionsUITests.swift
 714	Sources/AppleScriptSupport.swift
-708	Sources/App/MenuBarExtraController.swift
+706	Sources/App/MenuBarExtraController.swift
 683	Sources/SocketControlSettings.swift
 664	Sources/App/ShortcutRoutingSupport.swift
 643	Sources/Panels/BrowserPopupWindowController.swift
